requirements.txt
httpx>=0.28.1,<0.29.0
numpy>=2.2.2,<3.0.0
pydantic-ai-slim[openai]>=0.0.23,<0.1.0
pydantic-settings>=2.7.1
typer>=0.15.1,<0.16.0
pytest>=8.3.4,<9.0.0
pytest-asyncio>=0.25.3,<0.26.0